Внедрил рекомендации наставника по module2-task1
