+++ js/main.js
@@ -1,6 +1,6 @@
-import renderScreen from './app.js';
+import {renderScreen} from './render-screen.js';
 
-import screenIntro from './screens/screen-1.js';
+import screenIntro from './screens/screen-intro.js';
 
 // отрисовываем первый экран
 renderScreen(screenIntro);
