+++ js/screens/screen-game-1.js
@@ -1,13 +1,13 @@
-import {getElementFromTemplate} from '../model.js';
+//import {renderScreen, getElementFromTemplate} from '../model.js';
 
-import renderScreen from '../app.js';
+import {getElementFromTemplate, renderScreen} from '../render-screen.js';
 
-import screenGame3 from './screen-6.js';
+import screenGame2 from './screen-game-2.js';
 
-import screenGreeting from './screen-2.js';
+import screenGreeting from './screen-greeting.js';
 
 /**
- * 5. Экран второй игры, блок #game-2
+ * 4. Экран первой игры, блок #game-1
  * @type {Element}
  */
 const element = getElementFromTemplate(`
@@ -26,16 +26,27 @@ const element = getElementFromTemplate(`
       </div>
     </header>
     <div class="game">
-      <p class="game__task">Угадай, фото или рисунок?</p>
-      <form class="game__content  game__content--wide">
+      <p class="game__task">Угадайте для каждого изображения фото или рисунок?</p>
+      <form class="game__content">
         <div class="game__option">
-          <img src="http://placehold.it/705x455" alt="Option 1" width="705" height="455">
-          <label class="game__answer  game__answer--photo">
+          <img src="http://placehold.it/468x458" alt="Option 1" width="468" height="458">
+          <label class="game__answer game__answer--photo">
             <input name="question1" type="radio" value="photo">
             <span>Фото</span>
           </label>
-          <label class="game__answer  game__answer--wide  game__answer--paint">
+          <label class="game__answer game__answer--paint">
             <input name="question1" type="radio" value="paint">
+            <span>Рисунок</span>
+          </label>
+        </div>
+        <div class="game__option">
+          <img src="http://placehold.it/468x458" alt="Option 2" width="468" height="458">
+          <label class="game__answer  game__answer--photo">
+            <input name="question2" type="radio" value="photo">
+            <span>Фото</span>
+          </label>
+          <label class="game__answer  game__answer--paint">
+            <input name="question2" type="radio" value="paint">
             <span>Рисунок</span>
           </label>
         </div>
@@ -46,11 +57,11 @@ const element = getElementFromTemplate(`
           <li class="stats__result stats__result--slow"></li>
           <li class="stats__result stats__result--fast"></li>
           <li class="stats__result stats__result--correct"></li>
-          <li class="stats__result stats__result--wrong"></li>
           <li class="stats__result stats__result--unknown"></li>
-          <li class="stats__result stats__result--slow"></li>
           <li class="stats__result stats__result--unknown"></li>
-          <li class="stats__result stats__result--fast"></li>
+          <li class="stats__result stats__result--unknown"></li>
+          <li class="stats__result stats__result--unknown"></li>
+          <li class="stats__result stats__result--unknown"></li>
           <li class="stats__result stats__result--unknown"></li>
         </ul>
       </div>
@@ -66,15 +77,16 @@ const element = getElementFromTemplate(`
       </div>
     </footer>`);
 
-// Экран третьей игры, блок #game-3, должен показываться
-// по нажатию на любой ответ на втором игровом экране, любой блок .game__answer.
+
+// Экран второй игры, блок #game-2, должен показываться после того,
+// как будут выбраны оба варианта ответа на экране #game-1.
 const form = element.querySelector(`.game__content`);
+const options = element.querySelectorAll(`.game__option`);
 
 form.addEventListener(`change`, () => {
-  const gameAnswers = form.querySelectorAll(`.game__answer input:checked`);
-
-  if (gameAnswers.length >= 1) {
-    renderScreen(screenGame3);
+  const answers = form.querySelectorAll(`input:checked`);
+  if (options.length === answers.length) {
+    renderScreen(screenGame2);
   }
 });
