+++ js/screens/screen-game-2.js
@@ -1,13 +1,13 @@
-import {getElementFromTemplate} from '../model.js';
+//import {getElementFromTemplate} from '../model.js';
 
-import renderScreen from '../app.js';
+import {getElementFromTemplate, renderScreen} from '../render-screen.js';
 
-import screenGame2 from './screen-5.js';
+import screenGame3 from './screen-game-3.js';
 
-import screenGreeting from './screen-2.js';
+import screenGreeting from './screen-greeting.js';
 
 /**
- * 4. Экран первой игры, блок #game-1
+ * 5. Экран второй игры, блок #game-2
  * @type {Element}
  */
 const element = getElementFromTemplate(`
@@ -26,27 +26,16 @@ const element = getElementFromTemplate(`
       </div>
     </header>
     <div class="game">
-      <p class="game__task">Угадайте для каждого изображения фото или рисунок?</p>
-      <form class="game__content">
+      <p class="game__task">Угадай, фото или рисунок?</p>
+      <form class="game__content  game__content--wide">
         <div class="game__option">
-          <img src="http://placehold.it/468x458" alt="Option 1" width="468" height="458">
-          <label class="game__answer game__answer--photo">
+          <img src="http://placehold.it/705x455" alt="Option 1" width="705" height="455">
+          <label class="game__answer  game__answer--photo">
             <input name="question1" type="radio" value="photo">
             <span>Фото</span>
           </label>
-          <label class="game__answer game__answer--paint">
+          <label class="game__answer  game__answer--wide  game__answer--paint">
             <input name="question1" type="radio" value="paint">
-            <span>Рисунок</span>
-          </label>
-        </div>
-        <div class="game__option">
-          <img src="http://placehold.it/468x458" alt="Option 2" width="468" height="458">
-          <label class="game__answer  game__answer--photo">
-            <input name="question2" type="radio" value="photo">
-            <span>Фото</span>
-          </label>
-          <label class="game__answer  game__answer--paint">
-            <input name="question2" type="radio" value="paint">
             <span>Рисунок</span>
           </label>
         </div>
@@ -57,11 +46,11 @@ const element = getElementFromTemplate(`
           <li class="stats__result stats__result--slow"></li>
           <li class="stats__result stats__result--fast"></li>
           <li class="stats__result stats__result--correct"></li>
+          <li class="stats__result stats__result--wrong"></li>
           <li class="stats__result stats__result--unknown"></li>
+          <li class="stats__result stats__result--slow"></li>
           <li class="stats__result stats__result--unknown"></li>
-          <li class="stats__result stats__result--unknown"></li>
-          <li class="stats__result stats__result--unknown"></li>
-          <li class="stats__result stats__result--unknown"></li>
+          <li class="stats__result stats__result--fast"></li>
           <li class="stats__result stats__result--unknown"></li>
         </ul>
       </div>
@@ -77,17 +66,15 @@ const element = getElementFromTemplate(`
       </div>
     </footer>`);
 
-
-// Экран второй игры, блок #game-2, должен показываться после того,
-// как будут выбраны оба варианта ответа на экране #game-1.
+// Экран третьей игры, блок #game-3, должен показываться
+// по нажатию на любой ответ на втором игровом экране, любой блок .game__answer.
 const form = element.querySelector(`.game__content`);
 
 form.addEventListener(`change`, () => {
-  const options = form.querySelectorAll(`.game__option`);
-  const answers = form.querySelectorAll(`input:checked`);
+  const gameAnswers = form.querySelectorAll(`.game__answer input:checked`);
 
-  if (options.length === answers.length) {
-    renderScreen(screenGame2);
+  if (gameAnswers.length >= 1) {
+    renderScreen(screenGame3);
   }
 });
